Fixed hundreds of errors and warnings given by Eclipse including: Removed unused code/variables, Removed catch for exceptions that are not thrown, Removed unused import statements, Removed unnecessary casts,  Commenting out unused files, removing unnecessary semicolons, static references are made through class name not an object of the class

+++ projections/gui/TimelineAxisCanvas.java
@@ -114,8 +114,8 @@ public class TimelineAxisCanvas extends Canvas
 			break;
 		    }
 		    if (x % data.labelIncrement == 0) {  
-			tmp = format_.format((long)(data.beginTime + 
-						    (long)labeloffset + 
+			tmp = format_.format((data.beginTime + 
+						    labeloffset + 
 						    (long)x*data.timeIncrement));
 			og.drawLine(curx, axispos-5, curx, axispos + 5);
 			og.drawString(tmp, curx - fm.stringWidth(tmp)/2,
